add Vesting Fees withdraw in staking page.
