+++ src/app/hooks/staking/useStaking_getAccumulatedFees.ts
@@ -1,9 +1,13 @@
 import { useCacheCallWithValue } from '../useCacheCallWithValue';
 import { ethGenesisAddress } from '../../../utils/classifiers';
 
-export function useStaking_getAccumulatedFees(address: string, token: string) {
+export function useStaking_getAccumulatedFees(
+  address: string,
+  token: string,
+  useNewContract = false,
+) {
   return useCacheCallWithValue(
-    'feeSharingProxy_old',
+    useNewContract ? 'feeSharingProxy' : 'feeSharingProxy_old',
     'getAccumulatedFees',
     !!address && address !== ethGenesisAddress,
     address || ethGenesisAddress,
